refactor: split work-tile candidacy prefilter for 500-line gate (#3877)
Extract prefilter dispatch into order_suggestion_work_tile_candidacy_prefilter.dart
so repo.domain_package_source_file_size passes on the unified candidacy module.
